Move FunctionDescriptor into the polymer element descriptor file
