+++ src/ast/ast.ts
@@ -17,7 +17,6 @@ export * from './document-descriptor';
 export {ElementDescriptor, Attribute} from './element-descriptor';
 export * from './event-descriptor';
 export * from './feature-descriptor';
-export * from './function-descriptor';
 export * from './import-descriptor';
 export * from './inline-document-descriptor';
 export * from './property-descriptor';
