+++ src/editor-service.ts
@@ -16,7 +16,7 @@ import * as parse5 from 'parse5';
 
 import {Analysis} from './analysis';
 import {Analyzer} from './analyzer';
-import {DocumentDescriptor, FunctionDescriptor, Property} from './ast/ast';
+import {DocumentDescriptor, Property} from './ast/ast';
 import {SourceLocation} from './elements-format';
 import {HtmlDocument} from './html/html-document';
 import {PolymerElementDescriptor} from './polymer/element-descriptor';
